nivel-04/continuando-back-end/src/modules/appointments/services/CreateAppointmentService.ts
import { startOfHour } from 'date-fns';

import { injectable, inject } from 'tsyringe';

import AppError from '@shared/errors/AppError';
import Appointment from '../infra/typeorm/entities/Appointment';

import IAppointmentsRepository from '../repositories/IAppointmentsRepository';;

interface IRequest {
    provider_id: string
    user_id: string;
    date: Date;
}

@injectable()
class CreateAppointmentService {

    constructor(
        @inject('AppointmentsRepository')
        private appointmentsRepository: IAppointmentsRepository,

    ) {}

    public async execute({ provider_id, date, user_id }: IRequest): Promise<Appointment> {

        const appointmentDate = startOfHour(date);

        // Verificação da disponibilidade da data escolhida.
        const findAppointmentInSameDate = await this.appointmentsRepository.findByDate(
            appointmentDate,
        );

        // Retorna um erro caso a data já tenha sido escolhida.
        if (findAppointmentInSameDate) {
            throw new AppError('This appointment is already booked!');
        }

        const appointment = await this.appointmentsRepository.create({
            provider_id,
            user_id,
            date: appointmentDate,
        });

        return appointment;
    }
}

export default CreateAppointmentService;
